ENH: Select foreground mask in otsu
Otsu thresholding would often assign the mask value to the background of
the image, rather than the object within the image.

This commit looks at the corner (0,0,0) pixel in the otsu mask, and if
it isn't 0, swaps the otsu mask values so that it is zero.

This assumes the 0,0,0 pixel should be the background in the maask.